refactor: 탑승객 및 탑승객 dto에 이름 필드 추가
