+++ reservation-service/src/main/java/com/oneul_tanda/reservation_service/reservation/application/command/ConfirmReservationCommand.java
@@ -34,12 +34,13 @@ public record ConfirmReservationCommand(
 
 
     public record ConfirmPassengerCommand(
+            String name,
             String birth,
             Gender gender,
             String passportNumber
     ) {
         public static ConfirmPassengerCommand from(ConfirmReservationRequestDto.ConfirmPassengerDto dto) {
-            return new ConfirmPassengerCommand(dto.birth(), dto.gender(), dto.passportNumber());
+            return new ConfirmPassengerCommand(dto.name(), dto.birth(), dto.gender(), dto.passportNumber());
         }
     }
 }
