+++ reservation-service/src/main/java/com/oneul_tanda/reservation_service/reservation/presentation/dto/request/update/ConfirmReservationRequestDto.java
@@ -15,6 +15,7 @@ public record ConfirmReservationRequestDto(
     ) {}
 
     public record ConfirmPassengerDto(
+            String name,
             String birth,
             Gender gender,
             String passportNumber
